@@ -1,3 +1,3 @@
 module day1
 
-go 1.21.4
+go 1.25.4
